refactor!: refactor proto-actor project structure
This is the first step of the refactoring to update the Proto Actor Kotlin version to match the latest C# one.

BREAKING CHANGE: The project organization mimics the C# version, so some projects where merged. Also, there is some functionality not ported yet and a lot of code to update and convert. The Java support was removed for now, and the project heavily uses Kotlin coroutines.